img2imgservice/requirements.txt
fastapi
uvicorn
torch
diffusers[torch]
transformers
pillow
python-multipart